Moved echelon to another module and extracted nested functions.
