+++ src/matrices.rs
@@ -14,12 +14,6 @@ pub struct Matrix<T: MatrixNumber> {
     data: Vec<Vec<T>>,
 }
 
-#[derive(Debug, Clone)]
-pub struct Aftermath<T: MatrixNumber> {
-    result: Matrix<T>,
-    steps: Vec<String>,
-}
-
 impl<T: MatrixNumber> Matrix<T> {
     /// Creates a new matrix from a 2D vector.
     /// The matrix is not checked for validity.
@@ -160,126 +154,6 @@ impl<T: MatrixNumber> Matrix<T> {
         Self::new_unsafe(vec![])
     }
 
-    /// TODO: doc
-    /// TODO: maybe move to matrix_algorithm.rs
-    pub fn echelon(mut self) -> anyhow::Result<Aftermath<T>> {
-        const CONTEXT: &str = "Calculations error!";
-
-        if self.data.is_empty() {
-            return Ok(Aftermath::<T> {
-                result: self,
-                steps: Vec::new(),
-            });
-        }
-
-        let rows = self.data.len();
-        let cols = self.data[0].len();
-
-        let mut steps = vec![self.to_latex()];
-        let mut c = 0;
-        let mut i = 0;
-
-        let mut add_step = |step: &str, matrix: &Matrix<T>| {
-            steps.push(format!(r"\xrightarrow{{{}}} {}", step, matrix.to_latex()));
-        };
-
-        fn nice<T: MatrixNumber>(coef: &T) -> Option<i64> {
-            if coef.is_zero() {
-                Some(1000)
-            } else if coef.is_one() {
-                Some(0)
-            } else if (T::zero().checked_sub(coef)?).is_one() {
-                Some(1)
-            } else {
-                Some(2)
-            }
-        }
-
-        fn sub_coeff_to_latex<T: MatrixNumber>(coef: &T) -> Option<String> {
-            if coef.is_one() {
-                Some("- ".to_string())
-            } else if (T::zero().checked_sub(coef)?).is_one() {
-                Some("+ ".to_string())
-            } else if coef.is_positive() {
-                Some(format!("- {}", coef.to_latex()))
-            } else if coef.is_negative() {
-                Some(format!("+ {}", (T::zero().checked_sub(coef)?.to_latex())))
-            } else {
-                unreachable!("Should not be used for zero coefficient!")
-            }
-        }
-
-        while c < cols && i < rows {
-            let mut j = i;
-            for k in i + 1..rows {
-                if nice(&self.data[k][c]).context(CONTEXT)?
-                    < nice(&self.data[j][c]).context(CONTEXT)?
-                {
-                    j = k;
-                }
-            }
-
-            if !self.data[j][c].is_zero() {
-                if i != j {
-                    self.data.swap(i, j);
-                    add_step(
-                        format!(r"w_{{{}}} \leftrightarrow w_{{{}}}", i + 1, j + 1).as_str(),
-                        &self,
-                    );
-                }
-
-                if !self.data[i][c].is_one() {
-                    let d = self.data[i][c].clone();
-                    for j in c..cols {
-                        self.data[i][j] = self.data[i][j].checked_div(&d).context(CONTEXT)?;
-                    }
-
-                    add_step(
-                        format!(r"w_{{{}}} : {}", i + 1, d.to_latex_single()).as_str(),
-                        &self,
-                    );
-                }
-
-                let mut step_ops: Vec<String> = Vec::new();
-                for j in 0..rows {
-                    if j != i && !self.data[j][c].is_zero() {
-                        let p = self.data[j][c]
-                            .checked_div(&self.data[i][c])
-                            .context(CONTEXT)?;
-                        for k in c..cols {
-                            self.data[j][k] = self.data[j][k]
-                                .checked_sub(&self.data[i][k].checked_mul(&p).context(CONTEXT)?)
-                                .context(CONTEXT)?;
-                        }
-
-                        step_ops.push(format!(
-                            "w_{{{}}} {}w_{{{}}}",
-                            j + 1,
-                            sub_coeff_to_latex(&p).context(CONTEXT)?,
-                            i + 1
-                        ));
-                    }
-                }
-
-                if !step_ops.is_empty() {
-                    add_step(
-                        format!(r"\substack{{{}}}", &step_ops.join(r"\\")).as_str(),
-                        &self,
-                    );
-                }
-
-                i += 1;
-            }
-
-            c += 1;
-        }
-
-        Ok(Aftermath {
-            result: self,
-            steps,
-        })
-    }
-
     /// Returns the shape of the matrix.
     /// If the matrix is not valid, the behavior is undefined.
     /// If the matrix is empty, the shape is (0, 0).
@@ -337,6 +211,28 @@ impl<T: MatrixNumber> Matrix<T> {
                 })
                 .0 // does any row have different length?
         };
+    }
+
+    /// Returns the raw data of the matrix.
+    /// # Examples
+    /// ```rust
+    /// use matrix::Matrix;
+    /// let m = Matrix::new(vec![vec![1, 2], vec![3, 4]]).unwrap();
+    /// assert_eq!(m.get_data(), &vec![vec![1, 2], vec![3, 4]]);
+    /// ```
+    pub fn get_data(&self) -> &Vec<Vec<T>> {
+        &self.data
+    }
+
+    /// Returns the raw data of the matrix and consumes the matrix.
+    /// # Examples
+    /// ```rust
+    /// use matrix::Matrix;
+    /// let m = Matrix::new(vec![vec![1, 2], vec![3, 4]]).unwrap();
+    /// assert_eq!(m.into_data(), vec![vec![1, 2], vec![3, 4]]);
+    /// ```
+    pub fn consume(self) -> Vec<Vec<T>> {
+        self.data
     }
 
     /// Checks if two matrices have the same shape.
@@ -753,78 +649,6 @@ mod tests {
         assert_eq!(
             result.to_latex(),
             r"\begin{bmatrix}2 & 4 & 6\\8 & 10 & 12\end{bmatrix}"
-        );
-    }
-
-    #[test]
-    fn test_echelon_rational1() {
-        let m = rm![-2, 1; 1, 1];
-        let expected = rm![1, 0; 0, 1];
-
-        let aftermath = m.echelon().unwrap();
-
-        assert_eq!(aftermath.result.to_latex(), expected.to_latex());
-        assert_eq!(
-            aftermath.steps,
-            vec![
-                r"\begin{bmatrix}-2 & 1\\1 & 1\end{bmatrix}",
-                r"\xrightarrow{w_{1} \leftrightarrow w_{2}} \begin{bmatrix}1 & 1\\-2 & 1\end{bmatrix}",
-                r"\xrightarrow{\substack{w_{2} + 2w_{1}}} \begin{bmatrix}1 & 1\\0 & 3\end{bmatrix}",
-                r"\xrightarrow{w_{2} : 3} \begin{bmatrix}1 & 1\\0 & 1\end{bmatrix}",
-                r"\xrightarrow{\substack{w_{1} - w_{2}}} \begin{bmatrix}1 & 0\\0 & 1\end{bmatrix}",
-            ]
-        );
-    }
-
-    #[test]
-    fn test_echelon_rational2() {
-        let m = rm![4, 3; 2, 1];
-        let expected = rm![1, 0; 0, 1];
-
-        let aftermath = m.echelon().unwrap();
-
-        assert_eq!(aftermath.result.to_latex(), expected.to_latex());
-        assert_eq!(
-            aftermath.steps,
-            vec![
-                r"\begin{bmatrix}4 & 3\\2 & 1\end{bmatrix}",
-                r"\xrightarrow{w_{1} : 4} \begin{bmatrix}1 & \frac{3}{4}\\2 & 1\end{bmatrix}",
-                r"\xrightarrow{\substack{w_{2} - 2w_{1}}} \begin{bmatrix}1 & \frac{3}{4}\\0 & -\frac{1}{2}\end{bmatrix}",
-                r"\xrightarrow{w_{2} : \left(-\frac{1}{2}\right)} \begin{bmatrix}1 & \frac{3}{4}\\0 & 1\end{bmatrix}",
-                r"\xrightarrow{\substack{w_{1} - \frac{3}{4}w_{2}}} \begin{bmatrix}1 & 0\\0 & 1\end{bmatrix}",
-            ]
-        );
-    }
-
-    #[test]
-    fn test_echelon_rational3() {
-        let id = rm![1, 0; 0, 1];
-
-        let aftermath = id.clone().echelon().unwrap();
-
-        assert_eq!(aftermath.result.to_latex(), id.to_latex());
-        assert_eq!(
-            aftermath.steps,
-            vec![r"\begin{bmatrix}1 & 0\\0 & 1\end{bmatrix}"]
-        );
-    }
-
-    #[test]
-    fn test_echelon_rational4() {
-        let m = rm![1, -1, 1; 1, 1, -1; -1, 1, -1];
-        let expected = rm![1, 0, 0; 0, 1, -1; 0, 0, 0];
-
-        let aftermath = m.echelon().unwrap();
-
-        assert_eq!(aftermath.result.to_latex(), expected.to_latex());
-        assert_eq!(
-            aftermath.steps,
-            vec![
-                r"\begin{bmatrix}1 & -1 & 1\\1 & 1 & -1\\-1 & 1 & -1\end{bmatrix}",
-                r"\xrightarrow{\substack{w_{2} - w_{1}\\w_{3} + w_{1}}} \begin{bmatrix}1 & -1 & 1\\0 & 2 & -2\\0 & 0 & 0\end{bmatrix}",
-                r"\xrightarrow{w_{2} : 2} \begin{bmatrix}1 & -1 & 1\\0 & 1 & -1\\0 & 0 & 0\end{bmatrix}",
-                r"\xrightarrow{\substack{w_{1} + w_{2}}} \begin{bmatrix}1 & 0 & 0\\0 & 1 & -1\\0 & 0 & 0\end{bmatrix}",
-            ]
         );
     }
